@@ -15,7 +15,7 @@
         "source": "https://github.com/andrey-helldar/package-wizard"
     },
     "require": {
-        "php": "^7.3|^8.0",
+        "php": "^7.4|^8.0",
         "ext-json": "*",
         "composer-plugin-api": "^2.0",
         "andrey-helldar/support": "^2.6",
