@@ -3,22 +3,23 @@
 namespace Helldar\PackageWizard\Command;
 
 use Composer\Command\BaseCommand as ComposerBaseCommand;
-use Helldar\PackageWizard\Concerns\Input;
+use Helldar\PackageWizard\Concerns\Git;
+use Helldar\PackageWizard\Concerns\Output;
 use Helldar\PackageWizard\Concerns\Questionable;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\VarDumper\VarDumper;
 
 abstract class BaseCommand extends ComposerBaseCommand
 {
-    use Input;
+    use Git;
+    use Output;
     use Questionable;
 
     /** @var \Symfony\Component\Console\Input\InputInterface */
-    protected $input;
+    protected InputInterface $input;
 
     /** @var \Symfony\Component\Console\Output\OutputInterface */
-    protected $output;
+    protected OutputInterface $output;
 
     abstract public function handle();
 
@@ -28,14 +29,5 @@ abstract class BaseCommand extends ComposerBaseCommand
         $this->output = $output;
 
         $this->handle();
-    }
-
-    protected function basePath(): string
-    {
-        VarDumper::dump('aaaa');
-        $value = $this->getComposer(false)->getConfig()->get('data-dir');
-        VarDumper::dump('bbb: ' . $value);
-
-        return $value;
     }
 }
