@@ -3,14 +3,15 @@
 namespace Helldar\PackageWizard\Resources;
 
 use Helldar\Support\Concerns\Makeable;
+use Helldar\Support\Facades\Helpers\Str;
 
 final class License extends BaseResource
 {
     use Makeable;
 
-    protected $license;
+    protected string $license;
 
-    protected $authors = [];
+    protected array $authors = [];
 
     public function license(string $license): self
     {
@@ -30,13 +31,31 @@ final class License extends BaseResource
     {
         return $this->parser
             ->template($this->load())
-            ->replace('year', date('Y'))
-            ->replace('authors', implode(', ', $this->authors))
+            ->replace('license', $this->getLicense())
+            ->replace('year', $this->getYear())
+            ->replace('authors', $this->getAuthors())
             ->get();
+    }
+
+    protected function getLicense(): string
+    {
+        return $this->license;
+    }
+
+    protected function getYear(): int
+    {
+        return date('Y');
+    }
+
+    protected function getAuthors(): string
+    {
+        return implode(', ', $this->authors);
     }
 
     protected function path(): string
     {
-        return realpath(__DIR__ . '/../../resources/licenses/' . $this->license);
+        $filename = Str::lower($this->license);
+
+        return realpath(__DIR__ . '/../../resources/licenses/' . $filename);
     }
 }
