@@ -6,32 +6,27 @@ use Helldar\Support\Facades\Helpers\Str;
 
 final class ComposerPlugin extends BaseStepper
 {
-    protected $type = 'composer-plugin';
+    protected string $type = 'composer-plugin';
 
-    protected $require = [
+    protected array $require = [
         'php' => '^8.0',
 
         'composer-plugin-api' => '^2.0',
     ];
 
-    protected $require_dev = [
+    protected array $require_dev = [
         'composer/composer' => '^2.0',
         'mockery/mockery'   => '^1.0',
         'phpunit/phpunit'   => '^9.0',
         'symfony/console'   => '^5.0',
     ];
 
-    protected function fill(): void
+    public function getExtra(): array
     {
-        parent::fill();
+        $namespace = trim($this->getNamespace(), '/\\');
 
-        $this->fillExtra();
-    }
+        $namespace = Str::finish($namespace, '\\\\');
 
-    protected function fillExtra(): void
-    {
-        $namespace = Str::finish($this->getNamespace(), '\\');
-
-        $this->setExtra('class', $namespace . 'Application');
+        return ['class' => $namespace . 'Application'];
     }
 }
