@@ -4,11 +4,11 @@ namespace Helldar\PackageWizard\Steppers;
 
 final class Native extends BaseStepper
 {
-    protected $require = [
+    protected array $require = [
         'php' => '^8.0',
     ];
 
-    protected $require_dev = [
+    protected array $require_dev = [
         'mockery/mockery' => '^1.0',
         'phpunit/phpunit' => '^9.0',
     ];
