@@ -8,7 +8,7 @@ interface Stepperable extends Arrayable
 
     public function setName(string $name): void;
 
-    public function setDescription(?string $description): void;
+    public function setDescription(string $description): void;
 
     public function setLicense(string $license): void;
 
@@ -18,17 +18,15 @@ interface Stepperable extends Arrayable
 
     public function setAuthors(array $authors): void;
 
+    public function pushAuthor(array $author): void;
+
     public function getAuthors(): array;
 
-    public function setRequire(array $require): void;
+    public function setRequire(array $dependencies): void;
 
-    public function setRequireDev(array $require_dev): void;
-
-    public function setAutoload(): void;
+    public function setRequireDev(array $dev_dependencies): void;
 
     public function getAutoloadPath(): ?string;
-
-    public function setExtra(string $key, $value): void;
 
     public function setRepositoryUrl(string $url): void;
 }
