@@ -4,17 +4,19 @@ namespace Helldar\PackageWizard\Steps;
 
 use Helldar\Support\Facades\Helpers\Http;
 
-final class Url extends BaseStep
+final class RepositoryUrl extends BaseStep
 {
+    protected string $question = 'Repository URL of package: ';
+
     protected function input(): ?string
     {
-        $value = $this->io->ask($this->question);
+        $value = $this->getIO()->ask($this->question);
 
         if (Http::isUrl($value)) {
             return $value;
         }
 
-        $this->output->warning('An invalid URL was specified.');
+        $this->warning('An invalid URL was specified.');
 
         return $this->input();
     }
